ShippingMethodDraft to interface. #942
