+++ sphere-models/src/main/java/io/sphere/sdk/shippingmethods/ShippingMethodDraft.java
@@ -1,5 +1,6 @@
 package io.sphere.sdk.shippingmethods;
 
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import io.sphere.sdk.models.Reference;
 import io.sphere.sdk.models.Referenceable;
 import io.sphere.sdk.taxcategories.TaxCategory;
@@ -7,48 +8,25 @@ import io.sphere.sdk.taxcategories.TaxCategory;
 import javax.annotation.Nullable;
 import java.util.List;
 
-public class ShippingMethodDraft {
-    private final String name;
-    @Nullable
-    private final String description;
-    private final Reference<TaxCategory> taxCategory;
-    private final List<ZoneRate> zoneRates;
-    private final Boolean isDefault;
+@JsonDeserialize(as = ShippingMethodDraftImpl.class)
+public interface ShippingMethodDraft {
+    String getName();
 
-    private ShippingMethodDraft(final String name, @Nullable final String description, final Reference<TaxCategory> taxCategory, final List<ZoneRate> zoneRates, final Boolean isDefault) {
-        this.name = name;
-        this.description = description;
-        this.taxCategory = taxCategory;
-        this.zoneRates = zoneRates;
-        this.isDefault = isDefault;
-    }
+    @Nullable
+    String getDescription();
+
+    Reference<TaxCategory> getTaxCategory();
+
+    List<ZoneRate> getZoneRates();
+
+    Boolean isDefault();
+
 
     public static ShippingMethodDraft of(final String name, @Nullable final String description, final Referenceable<TaxCategory> taxCategory, final List<ZoneRate> zoneRates) {
         return of(name, description, taxCategory, zoneRates, false);
     }
 
     public static ShippingMethodDraft of(final String name, @Nullable final String description, final Referenceable<TaxCategory> taxCategory, final List<ZoneRate> zoneRates, final boolean isDefault) {
-        return new ShippingMethodDraft(name, description, taxCategory.toReference(), zoneRates, isDefault);
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    @Nullable
-    public String getDescription() {
-        return description;
-    }
-
-    public Reference<TaxCategory> getTaxCategory() {
-        return taxCategory;
-    }
-
-    public List<ZoneRate> getZoneRates() {
-        return zoneRates;
-    }
-
-    public Boolean isDefault() {
-        return isDefault;
+        return new ShippingMethodDraftImpl(name, description, taxCategory.toReference(), zoneRates, isDefault);
     }
 }
